requirements.txt
## The following requirements were added by pip freeze:
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==3.0
decorator==4.4.2
dependency-injector==3.30.0
pick==1.0.0
psycopg2-binary==2.8.5
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot==12.8
six==1.15.0
SQLAlchemy==1.3.18
tornado==6.0.4
